+++ Domain/Entities/Student.cs
@@ -5,9 +5,9 @@ public class Student : ObjectBase
 {
     public string? NameA { get; set; }
     public string? NameE { get; set; }
-    public int CountryId { get; set; }
+    public long CountryId { get; set; }
 
-    public int CityId { get; set; }
+    public long CityId { get; set; }
 
 
     public string? PhotoName { get; set; }
